Move HTTP client methods from BaseGitService to HTTPClient protocol
- Add concrete implementations of execute_request, handle_http_status_error, and handle_http_error to HTTPClient protocol
- Remove these methods from BaseGitService class
- Remove duplicate implementations from all mixin base classes (GitHubMixinBase, GitLabMixinBase, BitBucketMixinBase)
- Clean up imports in mixin classes
- Add provider property to HTTPClient protocol for error handling

This completes the transition from mixins to composition by centralizing HTTP client functionality in the protocol.
